+++ stroom-core/src/main/java/stroom/core/receive/DataStoreFileSetProcessor.java
@@ -24,6 +24,7 @@ import stroom.meta.api.StandardHeaderArguments;
 import stroom.meta.statistics.api.MetaStatistics;
 import stroom.proxy.repo.ErrorFileUtil;
 import stroom.proxy.repo.FileSet;
+import stroom.proxy.repo.FileSetKey;
 import stroom.proxy.repo.FileSetProcessor;
 import stroom.proxy.repo.ProxyFileHandler;
 import stroom.receive.common.StreamTargetStroomStreamHandler;
@@ -41,6 +42,7 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -82,19 +84,28 @@ public final class DataStoreFileSetProcessor implements FileSetProcessor {
         if (fileSet.getFiles().size() > 0) {
             final LogExecutionTime logExecutionTime = new LogExecutionTime();
 
-            final String feedName = fileSet.getFeed();
-            taskContextFactory.context("Processing set - " + feedName, taskContext -> {
-                LOGGER.info(() -> LogUtil.message("processFeedFiles() - Started {} ({} Files)",
-                        feedName, fileSet.getFiles().size()));
+            final FileSetKey key = fileSet.getKey();
+            taskContextFactory.context("Processing set - " + key, taskContext -> {
+                LOGGER.info(LogUtil.message("processFeedFiles() - Started {} ({} Files)", key, fileSet.getFiles().size()));
 
                 // Sort the files in the file set so there is some consistency to processing.
                 fileSet.getFiles().sort(Comparator.comparing(p -> p.getFileName().toString()));
-                LOGGER.debug(() -> LogUtil.message("process() - {} {}", feedName, fileSet.getFiles()));
+                LOGGER.debug(LogUtil.message("process() - {} {}", key, fileSet.getFiles()));
+
+                final String feedName = key.getFeedName();
+
+                String typeName = Optional.ofNullable(key.getTypeName())
+                        .map(String::trim)
+                        .orElse("");
+                if (typeName.isEmpty()) {
+                    // Get the default type name for this feed if none has been provided.
+                    typeName = feedProperties.getStreamTypeName(feedName);
+                }
 
                 // We don't want to aggregate reference feeds.
                 final boolean oneByOne = feedProperties.isReference(feedName) || !aggregate;
 
-                List<StreamTargetStroomStreamHandler> handlers = openStreamHandlers(feedName);
+                List<StreamTargetStroomStreamHandler> handlers = openStreamHandlers(feedName, typeName, oneByOne);
                 List<Path> deleteFileList = new ArrayList<>();
 
                 long sequence = 1;
@@ -120,7 +131,7 @@ public final class DataStoreFileSetProcessor implements FileSetProcessor {
 
                             // Start new batch
                             deleteFileList = new ArrayList<>();
-                            handlers = openStreamHandlers(feedName);
+                            handlers = openStreamHandlers(feedName, typeName, oneByOne);
                             sequence = 1;
                         }
 
@@ -134,23 +145,18 @@ public final class DataStoreFileSetProcessor implements FileSetProcessor {
                 }
                 closeStreamHandlers(handlers);
                 cleanup(deleteFileList);
-                LOGGER.info(() -> LogUtil.message("processFeedFiles() - Completed {} in {}",
-                        feedName, logExecutionTime));
+                LOGGER.info(LogUtil.message("processFeedFiles() - Completed {} in {}", feedName, logExecutionTime));
             }).run();
         }
     }
 
-    private List<StreamTargetStroomStreamHandler> openStreamHandlers(final String feedName) {
-        // We don't want to aggregate reference feeds.
-        final boolean oneByOne = feedProperties.isReference(feedName) || !aggregate;
-
+    private List<StreamTargetStroomStreamHandler> openStreamHandlers(final String feedName, final String typeName, final boolean oneByOne) {
         final StreamTargetStroomStreamHandler streamTargetStroomStreamHandler = new StreamTargetStroomStreamHandler(store,
-                feedProperties, metaStatistics, feedName, feedProperties.getStreamTypeName(feedName));
-
-        streamTargetStroomStreamHandler.setOneByOne(oneByOne);
+                feedProperties, metaStatistics, feedName, typeName, oneByOne);
 
         final AttributeMap globalMetaMap = new AttributeMap();
         globalMetaMap.put(StandardHeaderArguments.FEED, feedName);
+        globalMetaMap.put(StandardHeaderArguments.TYPE, typeName);
 
         streamTargetStroomStreamHandler.handleHeader(globalMetaMap);