+++ stroom-meta/stroom-meta-impl-db/src/main/java/stroom/meta/impl/db/MetaTypeDaoImpl.java
@@ -19,8 +19,8 @@ package stroom.meta.impl.db;
 
 import stroom.cache.api.CacheManager;
 import stroom.cache.api.ICache;
-import stroom.data.shared.StreamTypeNames;
 import stroom.db.util.JooqUtil;
+import stroom.meta.impl.MetaServiceConfig;
 import stroom.meta.impl.MetaTypeDao;
 import stroom.meta.impl.db.jooq.tables.records.MetaTypeRecord;
 import stroom.util.shared.Clearable;
@@ -30,6 +30,7 @@ import org.jooq.Field;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -50,12 +51,14 @@ class MetaTypeDaoImpl implements MetaTypeDao, Clearable {
         cache = cacheManager.create(CACHE_NAME, metaServiceConfig::getMetaTypeCache, this::load);
 
         // Ensure some types are preloaded.
-        load(StreamTypeNames.RAW_EVENTS);
-        load(StreamTypeNames.RAW_REFERENCE);
-        load(StreamTypeNames.EVENTS);
-        load(StreamTypeNames.REFERENCE);
-        load(StreamTypeNames.RECORDS);
-        load(StreamTypeNames.ERROR);
+        final String metaTypes = metaServiceConfig.getMetaTypes();
+        if (metaTypes != null && !metaTypes.isEmpty()) {
+            Arrays
+                    .stream(metaTypes.split("\n"))
+                    .map(String::trim)
+                    .filter(s -> !s.isEmpty())
+                    .forEach(this::load);
+        }
     }
 
     private int load(final String name) {