+++ stroom-proxy/stroom-proxy-repo/src/main/java/stroom/proxy/repo/RepositoryProcessor.java
@@ -284,7 +284,7 @@ public final class RepositoryProcessor {
         private final TaskContext parentContext;
         private final TaskContextFactory taskContextFactory;
 
-        private final Map<String, FileSet> fileSetMap = new ConcurrentHashMap<>();
+        private final Map<FileSetKey, FileSet> fileSetMap = new ConcurrentHashMap<>();
         private final Set<CompletableFuture<Void>> futures = Collections.newSetFromMap(new ConcurrentHashMap<>());
 
         private int totalMappedFiles;
@@ -317,14 +317,14 @@ public final class RepositoryProcessor {
 
         @Override
         public synchronized void accept(final ZipInfo zipInfo) {
-            final String feedName = zipInfo.getFeedName();
-            if (feedName == null || feedName.length() == 0) {
+            final FileSetKey key = zipInfo.getKey();
+            if (key.getFeedName() == null || key.getFeedName().isBlank()) {
                 errorReceiver.onError(zipInfo.getPath(), "Unable to find feed in header??");
 
             } else {
-                LOGGER.debug("{} belongs to feed {}", zipInfo.getPath(), feedName);
+                LOGGER.debug("{} belongs to {}", zipInfo.getPath(), key);
 
-                FileSet fileSet = fileSetMap.computeIfAbsent(feedName, k -> new FileSet(feedName));
+                FileSet fileSet = fileSetMap.computeIfAbsent(key, FileSet::new);
 
                 // See if the file set will overflow if we add this file.
                 if (fileSet.getFiles().size() > 0 &&
@@ -335,7 +335,7 @@ public final class RepositoryProcessor {
                     processFileSet(fileSet);
 
                     // Create a new file set.
-                    fileSet = fileSetMap.computeIfAbsent(feedName, k -> new FileSet(feedName));
+                    fileSet = fileSetMap.computeIfAbsent(key, FileSet::new);
                 }
 
                 // The file set is not full so add the file.
@@ -371,7 +371,7 @@ public final class RepositoryProcessor {
 
         private synchronized void processFileSet(final FileSet fileSet) {
             // Remove the full file set.
-            fileSetMap.remove(fileSet.getFeed());
+            fileSetMap.remove(fileSet.getKey());
 
             // Reduce the total number of files we have mapped.
             totalMappedFiles -= fileSet.getFiles().size();